nixos/tests/broken-nixos-subtest: init
This commit adds a new "broken-nixos-subtest" NixOS test as a proof
of concept of aborting interactive testing sessions when a subtest
fails.

The expected behaviour of this NixOS test is that the tester should be
brought back to the interactive python session in order to investigate a
failed subtest. Instead, the VM terminates entirely.

Steps to reproduce this behaviour:
1. nix build -f . nixosTests.broken-nixos-subtest.driverInteractive
2. ./result/bin/nixos-test-driver --interactive
3. test_script()